Cleanup ChunkReader (#4118) (#4156)
* Remove length from ChunkReader (#4118)

* Remove ChunkReader::T Send bound

* Remove FileSource

* Tweak docs
